requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0

SQLAlchemy==2.0.34
psycopg2-binary==2.9.9

passlib[bcrypt]==1.7.4
python-jose==3.3.0
tmdbsimple==2.9.1

pandas==2.2.2
numpy==2.0.1
scipy==1.14.1
scikit-learn==1.5.1
